Implement fully-connected network membership protocol
The gossip-style membership library used previously didn't seem to fully
implement the backing research paper and in practice the seed node would just
collect all peers without sending any identifiers back to any peers.

This protocol has peers join a seed node which returns the set of peers it
knows. The joining peer then performs a join operation for ach new peer in
this set so that every peer knows about all members.

This algorithm is super centralized but will work for this proof-of-concept.