Blocks miner to paper over a bug that will be fixed in future commit
As it stands with current development, nodes start mining on top of the genesis
block as soon as they launch. The mining process races with the peer discovery
process as they both use the scheduler in an unsynchronized way.

The codebase needs to be extended to reconcile "dump it all in the dispatch
loop" with "there are (now) tasks that need to block others upon launch".

The full change here probably touches the `scheduler`, the `queue` and the
`p2p` module as right now we assume an asynchronous model and to fix this bug
properly could require some substantive changes to take advantage of the
request/response semantics that are currently not exposed.